+++ library/playground.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import * as v from './dist/index.js';
+import * as v from './dist/index.mjs';