+++ website/src/components/ApiList.tsx
@@ -1,5 +1,5 @@
 import { component$ } from '@builder.io/qwik';
-import { Link } from '@builder.io/qwik-city';
+import { Link } from './Link';
 
 type ApiListProps = {
   label?: string;
@@ -10,11 +10,11 @@ type ApiListProps = {
  * List to display APIs and navigate to their documentation.
  */
 export const ApiList = component$<ApiListProps>(({ label, items }) => (
-  <ul class="ml-8! lg:ml-10! flex list-none flex-row flex-wrap gap-2">
+  <ul class="ml-8! flex list-none flex-row flex-wrap gap-2 lg:ml-10!">
     {label && label + ': '}
     {items.map((item, index) => (
       <li key={item} class="p-0!">
-        <Link href={`/api/${item}/`} prefetch={false}>
+        <Link href={`/api/${item}/`}>
           <code>{item}</code>
         </Link>
         {index < items.length - 1 && ','}