+++ website/src/components/CodeEditor.tsx
@@ -24,9 +24,9 @@ import prettierPluginEstree from 'prettier/plugins/estree';
 import prettierPluginTypeScript from 'prettier/plugins/typescript';
 import { formatWithCursor } from 'prettier/standalone';
 import { useTheme } from '~/routes/plugin@theme';
-import valibotTypes from '../../../library/dist/index.d.ts?raw';
+import valibotTypes from '../../../library/dist/index.d.mts?raw';
 import valibotPackageJson from '../../../library/package.json?raw';
-import valibotToJsonSchemaTypes from '../../../packages/to-json-schema/dist/index.d.ts?raw';
+import valibotToJsonSchemaTypes from '../../../packages/to-json-schema/dist/index.d.mts?raw';
 import valibotToJsonSchemaPackageJson from '../../../packages/to-json-schema/package.json?raw';
 import typescriptTm from '../json/TypeScript.tmLanguage.json';
 
@@ -403,31 +403,31 @@ async function setupMonaco() {
   });
 
   // Add Valibot to context of editor
-  monaco.languages.typescript.typescriptDefaults.addExtraLib(
+  monaco.typescript.typescriptDefaults.addExtraLib(
     valibotPackageJson,
     'file:///node_modules/valibot/package.json'
   );
-  monaco.languages.typescript.typescriptDefaults.addExtraLib(
+  monaco.typescript.typescriptDefaults.addExtraLib(
     valibotTypes,
-    'file:///node_modules/valibot/dist/index.d.ts'
+    'file:///node_modules/valibot/dist/index.d.mts'
   );
 
   // Add to-json-schema to context of editor
-  monaco.languages.typescript.typescriptDefaults.addExtraLib(
+  monaco.typescript.typescriptDefaults.addExtraLib(
     valibotToJsonSchemaPackageJson,
     'file:///node_modules/@valibot/to-json-schema/package.json'
   );
-  monaco.languages.typescript.typescriptDefaults.addExtraLib(
+  monaco.typescript.typescriptDefaults.addExtraLib(
     valibotToJsonSchemaTypes,
-    'file:///node_modules/@valibot/to-json-schema/dist/index.d.ts'
+    'file:///node_modules/@valibot/to-json-schema/dist/index.d.mts'
   );
 
   // Set TypeScript compiler options
-  monaco.languages.typescript.typescriptDefaults.setCompilerOptions({
+  monaco.typescript.typescriptDefaults.setCompilerOptions({
     strict: true,
-    target: monaco.languages.typescript.ScriptTarget.ESNext,
-    module: monaco.languages.typescript.ModuleKind.ESNext,
-    moduleResolution: monaco.languages.typescript.ModuleResolutionKind.NodeJs,
+    target: monaco.typescript.ScriptTarget.ESNext,
+    module: monaco.typescript.ModuleKind.ESNext,
+    moduleResolution: monaco.typescript.ModuleResolutionKind.NodeJs,
     skipLibCheck: true,
   });
 }