+++ website/src/components/Head.tsx
@@ -13,7 +13,6 @@ export const Head = component$(() => {
 
   // Compute document title
   const documentTitle = useComputed$(() =>
-    // eslint-disable-next-line qwik/valid-lexical-scope
     location.url.pathname === '/' ? head.title : `${head.title} | Valibot`
   );
 
@@ -31,7 +30,6 @@ export const Head = component$(() => {
 
   // Compute description from metadata
   const description = useComputed$(
-    // eslint-disable-next-line qwik/valid-lexical-scope
     () => head.meta.find((item) => item.name === 'description')?.content
   );
 
@@ -42,7 +40,6 @@ export const Head = component$(() => {
 
     // Add title and path to URL
     if (location.url.pathname !== '/') {
-      // eslint-disable-next-line qwik/valid-lexical-scope
       imageUrl += `?title=${encodeURIComponent(head.title)}&path=${
         location.url.pathname.split('/')[1]
       }`;