+++ website/src/components/UnstyledButton.tsx
@@ -1,5 +1,5 @@
 import { component$, type QRLEventHandlerMulti, Slot } from '@builder.io/qwik';
-import { Link } from '@builder.io/qwik-city';
+import { Link } from './Link';
 
 export type LinkButtonProps = {
   class?: string;
@@ -41,7 +41,7 @@ export const UnstyledButton = component$((props: UnstyledButtonProps) => {
 
     // Internal link
     return (
-      <Link {...props} prefetch={false}>
+      <Link {...props}>
         <Slot />
       </Link>
     );