+++ website/src/routes/playground/index.tsx
@@ -27,8 +27,8 @@ import {
 import { useResetSignal } from '~/hooks';
 import { BinIcon, CheckIcon, CopyIcon, PlayIcon, ShareIcon } from '~/icons';
 import { trackEvent } from '~/utils';
-import valibotCode from '../../../../library/dist/index.min.js?url';
-import valibotToJsonSchemaCode from '../../../../packages/to-json-schema/dist/index.min.js?url';
+import valibotCode from '../../../../library/dist/index.min.mjs?url';
+import valibotToJsonSchemaCode from '../../../../packages/to-json-schema/dist/index.min.mjs?url';
 import editorCode from './editorCode.ts?raw';
 import iframeCode from './iframeCode.js?raw';
 
@@ -171,7 +171,6 @@ export default component$(() => {
    * Captures logs from the iframe.
    */
   const captureLogs = $((event: MessageEvent<MessageEventData>) => {
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     if (event.data.type === 'log') {
       logs.value = [...logs.value, event.data.log];
     }
@@ -237,7 +236,7 @@ export default component$(() => {
           onSave$={saveCode}
         />
         <EditorButtons
-          class="hidden! lg:absolute! lg:flex! lg:right-10 lg:top-10 lg:z-10"
+          class="hidden! lg:absolute! lg:top-10 lg:right-10 lg:z-10 lg:flex!"
           model={model}
           executeCode$={executeCode}
         />
@@ -262,7 +261,7 @@ export default component$(() => {
           executeCode$={executeCode}
         />
         <IconButton
-          class="absolute! right-8 top-8 z-10 lg:right-10 lg:top-10"
+          class="absolute! top-8 right-8 z-10 lg:top-10 lg:right-10"
           type="button"
           variant="secondary"
           label="Clear logs"
@@ -368,7 +367,6 @@ const EditorButtons = component$<EditorButtonsProps>(
       const url = location.url.href;
 
       // Share URL or copy it to clipboard
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
       if (navigator.share) {
         navigator.share({ title: 'Playground', url });
       } else {